ima: refactoring of ima_init_policy()
This patch refactors the function ima_init_policy() by defining the new
function ima_array_add_rules() and the macro __add_rule_default().
Rules are added to the 'ima_default_rules' list depending on the value
of the 'ima_use_tcb' and 'ima_use_appraise_tcb' variables.

With this modification, it will be possible to write cleaner code for
loading a policy from multiple arrays of rules.

Signed-off-by: Roberto Sassu <rsassu@suse.de>