ima: added support for match_file policy rule criteria
This patch adds support for the new policy rule criteria 'match_file'.
An inode will be measured if the name of the accessed dentry will match
the pattern specified. At this stage, only regular expressions with the
'^' and '$' special characters are supported.

It is important to remark that this rule criteria can be easily
circumvented by renaming a file to prevent its measurement. Its use will
be primarily to prevent multiple measurements of the system libraries
(that match the pattern '.so$' or '.so.'). Measurement will be done in
the remaining cases by appending another rule without the 'match_file'
criteria.

Signed-off-by: Roberto Sassu <rsassu@suse.de>