ima: added new hook ima_bprm_committing_creds()
The existing hook ima_file_check() just inspects file accesses through
the open() system call. However, processes inheriting file descriptors
from their parent are not currently taken into account.

This patch adds the new hook ima_bprm_committing_creds(), which treats
the inheritance of file descriptors as the same as open() calls. With
this hook it will be possible to record the mapping of actions to every
process. Measurements through this hook are performed if the policy
contains a rule with the criteria 'func=INHERIT_FD_CHECK'.

The patch may be incomplete: when a file is appraised through this hook
(for example if a matching rule contains the subject label as criteria),
the return code of process_measurement() is not checked. To complete
the patch the same approach of SELinux could be followed: file descriptors
for which the appraising fails could be replaced with a new file
descriptor created for '/dev/null' (see flush_unauthorized_files()
in security/selinux/hooks.c for more details).

Changelog:
 - v1:
   - do not call process_measurement() from match_file() (Roberto Sassu)

Signed-off-by: Roberto Sassu <rsassu@suse.de>