ima: added new policy type read_exec
This patch defines the new policy type 'read_exec'. With this policy,
ima performs the same measurements of the 'exec' policy type and,
in addition, the reads and file descriptors inherited for each root
process. As for the 'exec' policy, shared libraries and also 'ld-so.cache'
are measured only once.

Signed-off-by: Roberto Sassu <rsassu@suse.de>