ima: do not set IMA_MEASURED if action matches policy rule with no_cache
This patch completes the implementation of the 'no_cache' policy option.
The current implementation suffers from the following issue: one can
circumvent the 'exec' or the 'read_exec' policies in order to hide the
loading of a shared library by one or multiple processes.

For example, the 'exec' policy matches events in the following order:
1) measure only once mmapped files that end with '.so', contain the '.so.'
   string and are owned by root
2) measure by process all other mmapped files

The root user may maliciously modify a system library and set a different
file owner. Then, he can measure the library and change the file owner
back to root. Since the IMA_MEASURED flag is set after the first
measurement, subsequent memory mapping operations will not be recorded
in the list (because they match the first rule).

With this patch, the attack described above will not be possible, because
the IMA_MEASURED flag will never be set on an inode if the action done
on it matches a rule with 'no_cache'. Thus, if the root user restores
the original system library owner, new accesses to the system library
will be recorded even if those actions match the first rule.

Merge with patch 'ima: added policy action flag no_cache'.